pyproject.toml
[project]
name = "memory-mcp"
version = "0.1.0"
description = "Local Memory MCP Server for OpenCode"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.7.4",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "openai>=2.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_mcp"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
